requirements.txt
discord.py>=2.0.0
pyyaml>=6.0.0
pandas>=1.5.0
numpy>=1.22.0
python-dotenv>=0.21.0
requests>=2.28.0
aiohttp>=3.8.3
structlog>=22.3.0
pytz>=2023.3
ibapi==9.81.1.post1
websocket-client>=1.4.0 